Rationalise utils in oml_utils and string_utils
