+++ lib/shared/headers.c
@@ -173,7 +173,7 @@
 
 #include "headers.h"
 #include "mem.h"
-#include "oml_util.h"
+#include "string_utils.h"
 
 static const struct {
   const char *name;
