+++ lib/shared/oml_utils.h
@@ -26,12 +26,6 @@
 
 #define LENGTH(a) ((sizeof (a)) / (sizeof ((a)[0])))
 
-void chomp (char* str);
-
-const char *skip_white (const char *p);
-const char *find_white (const char *p);
-const char *find_charn (const char *p, char c, int len);
-
 char* to_octets (unsigned char* buf, int len);
 
 int resolve_service(const char *service, int defport);
