+++ server/client_handler.c
@@ -25,8 +25,9 @@
 #include "mem.h"
 #include "monitoring_server.h"
 #include "mbuf.h"
+#include "string_utils.h"
 #include "oml_value.h"
-#include "oml_util.h"
+#include "oml_utils.h"
 #include "validate.h"
 #include "marshal.h"
 #include "binary.h"
