+++ server/database.c
@@ -58,7 +58,7 @@
 #include <assert.h>
 
 #include "ocomm/o_log.h"
-#include "oml_util.h"
+#include "oml_utils.h"
 #include "oml_value.h"
 #include "mem.h"
 #include "mstring.h"
