+++ test/lib/Makefile.am
@@ -11,8 +11,8 @@ AM_CPPFLAGS = \
 	-I  $(top_srcdir)/lib/shared
 
 check_liboml2_SOURCES = \
-	check_util.c \
-	check_util.h \
+	check_utils.c \
+	check_utils.h \
 	check_liboml2.c \
 	check_liboml2_api.c \
 	check_liboml2_bswap.c \
@@ -29,15 +29,15 @@ check_liboml2_SOURCES = \
 	$(top_srcdir)/lib/client/file_stream.h
 
 check_libshared_SOURCES = \
+	check_utils.c \
+	check_utils.h \
+	check_libshared.c \
 	check_libshared_base64.c \
 	check_libshared_json.c \
 	check_libshared_string_utils.c \
-	check_util.c \
-	check_util.h \
-	check_libshared.c \
 	check_libshared_suites.h \
 	check_libshared_mstring.c \
-	check_libshared_util.c \
+	check_libshared_oml_utils.c \
 	check_libshared_headers.c \
 	check_libshared_marshal.c
