+++ test/lib/check_liboml2_writers.c
@@ -31,7 +31,7 @@
 
 #include "mbuf.h"
 #include "client.h"
-#include "oml_util.h"
+#include "oml_utils.h"
 #include "file_stream.h"
 
 /*
