+++ test/lib/check_libshared_headers.c
@@ -5,7 +5,7 @@
 #include "headers.h"
 #include "text.h"
 #include "binary.h"
-#include "oml_util.h"
+#include "oml_utils.h"
 #include "oml_value.h"
 
 /*
