+++ test/server/check_binary_protocol.c
@@ -34,7 +34,7 @@
 #include "ocomm/o_log.h"
 #include "mem.h"
 #include "mbuf.h"
-#include "oml_util.h"
+#include "oml_utils.h"
 #include "oml_value.h"
 #include "marshal.h"
 #include "binary.h"
